requirements.txt
Flask==3.0.2
flask_mail==0.9.1
password_strength==0.0.3.post2
mysql-connector-python
python-dotenv==1.0.1
